+++ server/plugins/unifi/UniFiPlugin.ts
@@ -69,7 +69,7 @@ export class UniFiPlugin extends BasePlugin {
     private apiService: UniFiApiService;
 
     constructor() {
-        super('unifi', 'UniFi Controller', '0.7.2');
+        super('unifi', 'UniFi Controller', '0.7.3');
         this.apiService = new UniFiApiService();
     }